WIP: refactor enums in vsphere wizard
