+++ pkg/v1/tkg/web/src/app/views/landing/vsphere-wizard/vsphere-wizard.constants.ts
@@ -1,0 +1,20 @@
+import {NodeType} from "../wizard/shared/constants/wizard.constants";
+
+export const vSphereNodeTypes: Array<NodeType> = [
+    {
+        id: 'small',
+        name: 'small (cpu: 2, ram: 4 GB, disk: 20 GB)'
+    },
+    {
+        id: 'medium',
+        name: 'medium (cpu: 2, ram: 8 GB, disk: 40 GB)'
+    },
+    {
+        id: 'large',
+        name: 'large (cpu: 4, ram: 16 GB, disk: 40 GB)'
+    },
+    {
+        id: 'extra-large',
+        name: 'extra-large (cpu: 8, ram: 32 GB, disk: 80 GB)'
+    }
+];
